Moved pattern logic out of data catalog, implemented KedroDataCatalog
