+++ kedro/io/__init__.py
@@ -14,7 +14,7 @@ from .core import (
     Version,
 )
 from .data_catalog import DataCatalog
-from .data_catalog_redesign import AbstractDataCatalog, KedroDataCatalog
+from .data_catalog_redesign import KedroDataCatalog
 from .lambda_dataset import LambdaDataset
 from .memory_dataset import MemoryDataset
 from .shared_memory_dataset import SharedMemoryDataset
@@ -31,6 +31,5 @@ __all__ = [
     "MemoryDataset",
     "SharedMemoryDataset",
     "Version",
-    "AbstractDataCatalog",
     "KedroDataCatalog",
 ]
